:sparkles: Enable title coloring for Android (#101)
* show(): factor out no-param-reassign and FlowFixMe

The $FlowFixMe for processColor is no longer necessary because it is now properly typed in react-native. Additionally, disabling the no-param-reassign eslint rule is easily avoided, so this resolves that as well.

This shouldn't result in any behavioral changes.

* show(): Add optional color param for title text

The `color` option is processed the same way as the other color params and is then passed down to the native implementations where it can be used to style the snackbar's title text.

* Android: style title using color option if set

* README: Document color param

* Android displaySnackbar: remove duplicate calls

getView and findViewById were called multiple times, unnecessarily. This change should not result in any behavior changes.

* Minor cleanup

* Don't pass action object if none is specified

* Add text color to example with style

* Commit updated snapshots